+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+## v0.1.2 (2022-12-12)
 * Added publicDir config
 * Added isHot flag to manifest